+++ History.md
@@ -4,6 +4,7 @@ Unpublished
   * corrected initial release date
   * use callbacks rather than events for asychronous open method
   * documentation
+  * return this in synchronous write methods
 
 0.1.0 / Dec 09 2014
 ===================